Fix transformation matrix output
Previously, a 1 px border was added to the voxelized output, which
wasn't accounted for in the transformation matrix. Moreover, the STL
model is scaled to 1 px less than the resolution so that its edges fall
on the centers of the outer pixels, which wasn't accounted for when
calculating the resolution for a particular voxel spacing. This commit
fixes both of these issues.